Splitting Sankey references
